Platform check: Win -> linux, Android
This example can work on Windows if iTunes (which includes Apple Bonjour) is installed.  It cannot currently run on Linux (even with Avahi) or Android.
Note that this platform check still needs to be moved into the application's onLaunch before Zeroconf is initialized. If Bonjour is not installed, the app will break before the platform check is run.